requirements.txt
pyautogui
pillow
pynput
pytesseract
requests
openai
anthropic
google-genai
groq
transformers
torch
